Linux (embedded) driver for Ultrasonic sensor HC-SR04
Implementation of a proximity alert system based on STM32-F7 Discovery
board with uCLinux OS inside and the HC-SR04 Ultrasonic Ranging Module.
The board read the distance of any object located in front of the
HC-SR04 sensor, and then it actives the LED, which blinks depending on
the distance of the detected object.